Added log and fixed center bug
I added a logger.ahk file and implemented it throughout the code to create a log file that stores debug messages into a txt file, log.txt. I also made the center player function in player.ahk move the player closer to the center. The player kept getting stuck on the stairs because they were too close to the edge of the block. This does slow down the program a little.