Added an optional timer
Added an optional timer to start the harvest process at a specific time. There is a bug where if you put a start time that is one minute less than the current time it only waits for an hour instead of 20 hours, but I will fix this later. It's not a major bug. Otherwise this function works, as far as I know.