Added getCharNum method
Added a method called getCharNum to MinecraftWin.ahk. It is similar to getChar, but focuses only on reading numbers and numerical symbols, the period '.' and minus sign '-'.  Since the function only reads number it does not have to read every single pixel to differentiate between characters and runs much more quickly. The program can get the coordinate, pitch, and yaw about 75% quicker now.